feat: 공통 응답 DTO 추가

+++ core/src/main/kotlin/com/example/core/domain/product/dto/ProductResponse.kt
@@ -20,7 +20,7 @@ data class ProductResponse(
 
         fun from(product: ProductDocument): ProductResponse {
             return ProductResponse(
-                productId = product.id.toLong(),
+                productId = product.id,
                 name = product.name,
                 price = product.price
             )
